Big push.
Make the main interface functorized, which helps type check bother the
client and the server. `common.re` is a module that exposes a type and
a `stringify` function, which are used to make sure the client can't
listen on wrong messages etc...

Add support for most of the API (the easy parts at least).

Learnings from adhoc polymorphism:
http://bloomberg.github.io/bucklescript/Manual.html#_phantom_arguments_and_ad_hoc_polyrmophism

Problems related to typechecking external modules. This module works
when given an object http. This is really hard to typecheck because that
module, even if it might follow a precise interface, isn't accessible
from my code (without having an explicit dependency on it...?).