Big fix to example + support sending variants.
The big part of this commit is the new support for sending variants.
The previous commit showed what the problem was and for some reason I
decided to not tackle this issue inside resocketio.
Now it works! But I had to revert the crazy type magic I wanted to force
the user to handle all possible messages.